Apply fix for Feral Hosting setup
This patches _h5ai so it can work in both username.server.feralhosting.com and server.feralhosting.com/username patterns without erroneously redirecting when it loads up.

The first patch actually allows the info page to work on initial testing and the second one makes it so it generates directory listings correctly.

As far as I know nothing breaks.

This is based on a pre-existing patch that was already applied to a slightly modified 0.28.1 release and reapplied to the current 0.29.0 codebase.  Should work for the forseeable future unless this file is changed and this particular part of the logic altered.